+++ CHANGELOG.md
@@ -1,21 +1,50 @@
+# Change Log
+
+0.3.1
+
+* Build system improved
+* Support for Lakka and LibreElec
+* Self-extracting package system for systems without compiler
+* Issue #5: Fix install files to correct locations
+* Expand documentation of Shared memory IPC Protocol
+
+0.3.0
+
+* Rewrite of monitor_device function
+* Fix leak of File in log_message
+* Improved commenting
+* Implemented shared memory requests without using signals
+* Fix formatting errors in markdown files
+* PATCH for Argon Artik Fan hat stop fan on shutdown
+* Fix argonone-cli respect verbose and silent options
+* Change default LOGLEVEL to 5 LOG_INFO
+* Added new install script
+* Added install support for Arch Linux Arm
+
 0.2.2
+
 * PATCH Shared memory update after cooldown
 * Fix a typos in log output and function name
 * Confirmed Operation on the Argon One m.2 case
 * Confirmed Ubuntu is a fully working
 
 0.2.1
+
 * PATCH for Argon Artik Fan hat
 * Added OS support for Gentoo and Manjaro
 * Change Build Procedure for new OS support
 
 0.2.0
+
 * Added shared memory interface
 * Changed Logging system
-* Added CLI program to interface with daemon
- 
+* Added CLI program to interface with daemon  
 
 0.1.6  
+
 * fix signed unsigned conversion  
 * fix unexpected exit if pulse width out of bounds  
-0.1.5 Initial Release
+
+0.1.5
+
+* Initial Release